+++ src/types/blockmetadata.ts
@@ -1,20 +1,25 @@
 import LocalQuotes from "../main";
 import {parseCodeBlock} from "../util/parser";
-import {getRandomQuoteOfAuthor} from "../util/random";
 import {getBlockMetadataIdx} from "../util/scan";
 import {getCurrentSeconds} from "../util/date";
+import {searchQuote} from "./quote";
+
+export interface BlockMetadataContent {
+	author: string;
+	text: string;
+}
 
 export interface BlockMetadata {
 	id: string;
-	author: string;
-	text: string;
-	lastUpdate: number;
-	reloadInterval: number;
+	search: string;
+	content: BlockMetadataContent,
 	customClass: string;
+	refresh: number;
+	lastUpdate: number;
 }
 
 function makeBlockMetadata(plugin: LocalQuotes, rawBlockMetadata: BlockMetadata): BlockMetadata {
-	rawBlockMetadata.text = getRandomQuoteOfAuthor(plugin, rawBlockMetadata.author);
+	rawBlockMetadata.content = searchQuote(plugin, rawBlockMetadata.search);
 	rawBlockMetadata.lastUpdate = getCurrentSeconds();
 
 	plugin.settings.blockMetadata.push(rawBlockMetadata);
@@ -27,25 +32,25 @@ function updateBlockMetadata(plugin: LocalQuotes, rawBlockMetadata: BlockMetadat
 	const prevBm: BlockMetadata = plugin.settings.blockMetadata[bmIdx];
 
 	// Fields updating
-	if (prevBm.author !== rawBlockMetadata.author) {
-		plugin.settings.blockMetadata[bmIdx].author = rawBlockMetadata.author;
-		plugin.settings.blockMetadata[bmIdx].text = getRandomQuoteOfAuthor(plugin, rawBlockMetadata.author);
+	if (prevBm.search !== rawBlockMetadata.search) {
+		plugin.settings.blockMetadata[bmIdx].search = rawBlockMetadata.search;
+		plugin.settings.blockMetadata[bmIdx].content = searchQuote(plugin, rawBlockMetadata.search);
 	}
 	if (prevBm.customClass !== rawBlockMetadata.customClass) {
 		plugin.settings.blockMetadata[bmIdx].customClass = rawBlockMetadata.customClass;
 	}
-	if (prevBm.reloadInterval !== rawBlockMetadata.reloadInterval) {
-		plugin.settings.blockMetadata[bmIdx].reloadInterval = rawBlockMetadata.reloadInterval;
-		plugin.settings.blockMetadata[bmIdx].text = getRandomQuoteOfAuthor(plugin, rawBlockMetadata.author);
+	if (prevBm.refresh !== rawBlockMetadata.refresh) {
+		plugin.settings.blockMetadata[bmIdx].refresh = rawBlockMetadata.refresh;
+		plugin.settings.blockMetadata[bmIdx].content = searchQuote(plugin, rawBlockMetadata.search);
 	}
 
 	// Update quote
-	let reloadInterval = plugin.settings.blockMetadata[bmIdx].reloadInterval === null
+	let refreshInterval = plugin.settings.blockMetadata[bmIdx].refresh === null
 			? plugin.settings.defaultReloadInterval
-			: plugin.settings.blockMetadata[bmIdx].reloadInterval;
+			: plugin.settings.blockMetadata[bmIdx].refresh;
 
-	if ((plugin.settings.blockMetadata[bmIdx].lastUpdate + reloadInterval) < getCurrentSeconds()) {
-		plugin.settings.blockMetadata[bmIdx].text = getRandomQuoteOfAuthor(plugin, rawBlockMetadata.author);
+	if ((plugin.settings.blockMetadata[bmIdx].lastUpdate + refreshInterval) < getCurrentSeconds()) {
+		plugin.settings.blockMetadata[bmIdx].content = searchQuote(plugin, rawBlockMetadata.search);
 		plugin.settings.blockMetadata[bmIdx].lastUpdate = getCurrentSeconds();
 	}
 
@@ -56,14 +61,16 @@ export function selectBlockMetadata(plugin: LocalQuotes, source: string): BlockM
 	let tmpBm: BlockMetadata = parseCodeBlock(source);
 	const idx: number = plugin.settings.blockMetadata.findIndex((e) => e.id === tmpBm.id);
 
-	// If author and/or id aren't set
-	if (!(tmpBm.id && tmpBm.author) || !plugin.quoteVault.length) {
+	// If author and/or id aren't set, or quoteVault is empty
+	if (!(tmpBm.id && tmpBm.search) || !plugin.quoteVault.length) {
 		return {
-			author: 'Local Quotes',
-			text: 'You caught an error! Maybe you don\'t set author or/and id, maybe you\'ve no quotes in the vault.' +
-				' You always can write an issue on GitHub',
-			customClass: null, id: null, lastUpdate: null, reloadInterval: null
-		}
+			content: {
+				author: 'Local Quotes',
+				text: 'You caught an error! Maybe you don\'t set author or/and id, maybe you\'ve no quotes in the ' +
+					'vault. You always can write an issue on GitHub'
+			},
+			customClass: null, id: null, lastUpdate: 0, refresh: null, search: null
+		};
 	} else {
 		if (idx >= 0) {
 			return updateBlockMetadata(plugin, tmpBm);